Fork SC_License_Manager
